+++ mods/BrewLAN_Gameplay/Craters/mod_info.lua
@@ -1,0 +1,7 @@
+name = "Swiss Cheese"
+uid = "e696c83e-b131-446d-940f-0ae91adad552"
+version = 1
+copyright = "© 2022 Sean Wheeldon"
+description = "Damage causes craters."
+author = "Balthazar"
+--icon = "/mods/BrewLAN_Gameplay/Craters/mod_icon.dds"